update pass info
