+++ fangjia-boot-admin/src/main/resources/application.properties
@@ -5,11 +5,10 @@ eureka.client.serviceUrl.defaultZone=http://goojia:goojia123456@master:8761/eure
 eureka.instance.preferIpAddress=true
 eureka.instance.instance-id=${spring.application.name}:${spring.cloud.client.ipAddress}:${server.port}
 eureka.instance.status-page-url=http://${spring.cloud.client.ipAddress}:${server.port}
-management.security.enabled=false
 
 security.basic.enabled=true
-security.user.name=goojia
-security.user.password=goojia123456
+security.user.name=yinjihuan
+security.user.password=123456
 
 # turbine 服务名称
 spring.boot.admin.turbine.location=fangjia-hystrix-dashboard
