+++ fangjia-eureka/src/main/resources/application.properties
@@ -9,7 +9,7 @@ eureka.client.fetch-registry=false
 eureka.server.enableSelfPreservation=false
 
 security.basic.enabled=true
-security.user.name=goojia
-security.user.password=goojia123456
+security.user.name=yinjihuan
+security.user.password=123456
 
 spring.profiles.active=master
