+++ fangjia-fsh-api/src/main/resources/application.properties
@@ -1,6 +1,6 @@
 spring.application.name=fangjia-fsh-api
 server.port=2103
-eureka.client.serviceUrl.defaultZone=http://goojia:goojia123456@master:8761/eureka/
+eureka.client.serviceUrl.defaultZone=http://yinjihuan:123456@master:8761/eureka/
 eureka.instance.preferIpAddress=true
 eureka.instance.instance-id=${spring.application.name}:${spring.cloud.client.ipAddress}:${spring.application.instance_id:${server.port}}
