+++ fangjia-fsh-substitution-service/src/main/resources/application.properties
@@ -61,3 +61,6 @@ ribbon.maxAutoRetries=1
 ribbon.maxAutoRetriesNextServer=3
 # 对所有操作请求都进行重试
 ribbon.okToRetryOnAllOperations=true
+
+#zipkin 抽样比例
+spring.sleuth.sampler.percentage=1
