+++ fangjia-hystrix-dashboard/src/main/resources/application.properties
@@ -2,7 +2,7 @@ spring.application.name=fangjia-hystrix-dashboard
 
 server.port=9011
 
-eureka.client.serviceUrl.defaultZone=http://goojia:goojia123456@master:8761/eureka/
+eureka.client.serviceUrl.defaultZone=http://yinjihuan:123456@master:8761/eureka/
 eureka.instance.preferIpAddress=true
 eureka.instance.instance-id=${spring.application.name}:${spring.cloud.client.ipAddress}:${server.port}
 eureka.instance.status-page-url=http://${spring.cloud.client.ipAddress}:${server.port}/hystrix
